Depends on pycurl as it seems now more recent than pycurl2.
pycurl2 was a fork of pycurl as of 7.19.0 (Sep 9 2008).
As of 2015-08-10, last commit on this repo is : 2013-08-09 (2 years).
Now pycurl seems alive, current version being 7.19.5.1, with
last commit being 1 month ago.

As we introduce new constants absent from pycurl2, I think it's
better to depend on the original pycurl version.